feat: move Trip model from Core to DriverData package
